requirements.txt
arrow==0.12.1
psycopg2-binary==2.7.6.1
zeep==3.1.0

